Make SDL_URIToLocal available to multiple platforms
Moves the functions out of core/unix into SDL_utils.c
